+++ js/photo-data.js
@@ -1,0 +1,6 @@
+const PHOTO_COUNTER = 25;
+const PHOTO_DESCRIPTIONS = ['хорошее фото', 'нормальное фото', 'неудачное фото'];
+const COMMENT_MESSAGES = ['Всё отлично!','В целом всё неплохо. Но не всё.','Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.','Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.','Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.','Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+const USER_NAMES = ['Антон','Александр','Алексей','Бездомный','Валерий','Владимир','Валентин','Воланд'];
+
+export {PHOTO_COUNTER, PHOTO_DESCRIPTIONS, COMMENT_MESSAGES, USER_NAMES};
